Make use all atomic across tools
- prevent a later tool failure from leaving earlier agents switched
- share context snapshot and restore logic to keep both workflows aligned
